change sizes to int, fix resizing image logic

+++ src/main/java/xyz/melnychuk/niimbotprint/model/BarcodeElement.java
@@ -9,14 +9,14 @@ public class BarcodeElement extends StickerElement {
 
     private String content;
     private String format;
-    private double width;
-    private double height;
+    private int width;
+    private int height;
 
     public BarcodeElement() {
         this("1234567890", 0, 0, 200, 80, "CODE_128");
     }
 
-    public BarcodeElement(String content, double x, double y, double width, double height, String format) {
+    public BarcodeElement(String content, double x, double y, int width, int height, String format) {
         super(x, y);
         this.content = content;
         this.format = format;
